+++ lib/ui/index.js
@@ -3,14 +3,14 @@ const path = require('path');
 const { existsSync } = require('fs');
 
 let router = express.Router()
-  .use(express.static(path.join(richUiRoot, 'public')));
+  .use(express.static(path.join(__dirname, 'public')));
 
-if (existsSync(path.join(richUiRoot, 'oauth'))) {
-  router.use('/oauth', require(path.join(richUiRoot, 'oauth')));
+if (existsSync(path.join(__dirname, 'oauth'))) {
+  router.use('/oauth', require(path.join(__dirname, 'oauth')));
 }
 
-if (existsSync(path.join(richUiRoot, 'api'))) {
-  router.use(require(path.join(richUiRoot, 'api')));
+if (existsSync(path.join(__dirname, 'api'))) {
+  router.use(require(path.join(__dirname, 'api')));
 }
 
 module.exports = router;
