+++ lib/ui/public/app/router.js
@@ -1,8 +1,8 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import Config from './config/Config.vue';
-import PageOne from './config/PageOne.vue';
-import PageTwo from './config/PageTwo.vue';
+import Page1 from './config/Page1.vue';
+import Page2 from './config/Page2.vue';
 
 Vue.use(Router);
 
@@ -24,13 +24,13 @@ export default new Router({
         {
           path: '',
           alias: '1',
-          name: 'PageOne',
-          component: PageOne,
+          name: 'Page1',
+          component: Page1,
         },
         {
           path: '2',
-          name: 'PageTwo',
-          component: PageTwo,
+          name: 'Page2',
+          component: Page2,
         }
       ],
     },
